Move getDomainOfErrorBars to its own file
This is needed to avoid circular dependency between ErrorBar and axis selectors
