+++ scripts/snapshots/typesFiles.txt
@@ -18,6 +18,7 @@
   "types/util/isWellBehavedNumber.d.ts",
   "types/util/isDomainSpecifiedByUser.d.ts",
   "types/util/getEveryNthWithCondition.d.ts",
+  "types/util/getDomainOfErrorBars.d.ts",
   "types/util/cursor",
   "types/util/calculateViewBox.d.ts",
   "types/util/TickUtils.d.ts",
