+++ test/util/ChartUtils/getDomainOfErrorBars.spec.tsx
@@ -1,9 +1,9 @@
 import React, { ReactElement, ReactNode } from 'react';
 import { vi } from 'vitest';
 import { findAllByType } from '../../../src/util/ReactUtils';
-import { getDomainOfErrorBars } from '../../../src/util/ChartUtils';
 import { ErrorBar } from '../../../src/cartesian/ErrorBar';
 import { AxisType, LayoutType } from '../../../src/util/types';
+import { getDomainOfErrorBars } from '../../../src/util/getDomainOfErrorBars';
 
 vi.mock('../../../src/util/ReactUtils');
